Rename package name from "org.eclipse.dataspaceconnector" to "net.catenax.edc".
