+++ api-wrapper/services/api-wrapper/src/main/java/net/catenax/edc/apiwrapper/cache/InMemoryContractAgreementCache.java
@@ -1,0 +1,16 @@
+package net.catenax.edc.apiwrapper.cache;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class InMemoryContractAgreementCache {
+    private final Map<String, String> store = new HashMap<>();
+
+    public void put(String assetId, String contractAgreementId) {
+        store.put(assetId, contractAgreementId);
+    }
+
+    public String get(String assetId) {
+        return store.get(assetId);
+    }
+}
