+++ api-wrapper/services/api-wrapper/src/main/java/net/catenax/edc/apiwrapper/config/ApiWrapperConfigKeys.java
@@ -1,0 +1,20 @@
+package net.catenax.edc.apiwrapper.config;
+
+import org.eclipse.dataspaceconnector.spi.EdcSetting;
+
+public final class ApiWrapperConfigKeys {
+
+    @EdcSetting(required = true)
+    public static final String CONSUMER_EDC_DATAMANAGEMENT_URL = "wrapper.consumer.edc.datamanagement.url";
+
+    @EdcSetting
+    public static final String CONSUMER_EDC_APIKEY_NAME = "wrapper.consumer.edc.apikey.name";
+
+    @EdcSetting
+    public static final String CONSUMER_EDC_APIKEY_VALUE = "wrapper.consumer.edc.apikey.value";
+
+    @EdcSetting
+    public static final String BASIC_AUTH = "wrapper.auth.basic";
+
+    private ApiWrapperConfigKeys() {}
+}
