+++ api-wrapper/services/api-wrapper/src/main/java/net/catenax/edc/apiwrapper/connector/sdk/Utility.java
@@ -1,0 +1,9 @@
+package net.catenax.edc.apiwrapper.connector.sdk;
+
+import okhttp3.MediaType;
+
+public class Utility {
+    public static final MediaType JSON = MediaType.get("application/json");
+
+    private Utility() {}
+}
