+++ api-wrapper/services/api-wrapper/src/main/java/net/catenax/edc/apiwrapper/connector/sdk/model/NegotiationInitiateRequestDto.java
@@ -1,0 +1,83 @@
+/*
+ *  Copyright (c) 2020 - 2022 Microsoft Corporation
+ *
+ *  This program and the accompanying materials are made available under the
+ *  terms of the Apache License, Version 2.0 which is available at
+ *  https://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ *
+ *  Contributors:
+ *       Microsoft Corporation - initial API and implementation
+ *
+ */
+
+package net.catenax.edc.apiwrapper.connector.sdk.model;
+
+import org.jetbrains.annotations.NotNull;
+
+public class NegotiationInitiateRequestDto {
+
+    private String connectorAddress;
+    private String protocol = "ids-multipart";
+    private String connectorId;
+    private ContractOfferDescription offer;
+
+    private NegotiationInitiateRequestDto() {
+
+    }
+
+    public String getConnectorAddress() {
+        return connectorAddress;
+    }
+
+    public String getProtocol() {
+        return protocol;
+    }
+
+    public String getConnectorId() {
+        return connectorId;
+    }
+
+    public @NotNull ContractOfferDescription getOffer() {
+        return offer;
+    }
+
+
+    public static final class Builder {
+        private final NegotiationInitiateRequestDto dto;
+
+        private Builder() {
+            dto = new NegotiationInitiateRequestDto();
+        }
+
+        public static Builder newInstance() {
+            return new Builder();
+        }
+
+        public Builder connectorAddress(String connectorAddress) {
+            dto.connectorAddress = connectorAddress;
+            return this;
+        }
+
+        public Builder protocol(String protocol) {
+            dto.protocol = protocol;
+            return this;
+        }
+
+        public Builder connectorId(String connectorId) {
+            dto.connectorId = connectorId;
+            return this;
+        }
+
+        public Builder offerId(ContractOfferDescription offerId) {
+            dto.offer = offerId;
+            return this;
+        }
+
+        public NegotiationInitiateRequestDto build() {
+            return dto;
+        }
+    }
+}
+
