--- api-wrapper/services/api-wrapper/src/main/java/org/eclipse/dataspaceconnector/apiwrapper/connector/sdk/Utility.java
@@ -1,9 +1,0 @@
-package org.eclipse.dataspaceconnector.apiwrapper.connector.sdk;
-
-import okhttp3.MediaType;
-
-public class Utility {
-    public static final MediaType JSON = MediaType.get("application/json");
-
-    private Utility() {}
-}
