+++ api-wrapper/services/api-wrapper/src/main/resources/META-INF/services/org.eclipse.dataspaceconnector.spi.system.ServiceExtension
@@ -1,1 +1,1 @@
-org.eclipse.dataspaceconnector.apiwrapper.ApiWrapperExtension
+net.catenax.edc.apiwrapper.ApiWrapperExtension
